requirements.txt
browser_use
pydantic
langchain
langchain-openai
google-ai-generativelanguage
pytest
pytest-asyncio
pyyaml
pandas